Merge the output (it was split previously)
The initial implementation was intending to print the holiday's indication
also. Unfortunately the spreadsheet has blocked cells between holiday and
timesheet columns.

As workaround the output was split vertically. However, such solution demands
two copy-and-paste operations and the "region copy" feature supported on the
terminal (using CTRL modifier).

The solution brought by this patch try to simplify the output adding some extra
manual work (slightly in my opinion). The holiday indication (x) is being
printed together with the timesheet and demands manual edition on the
spreadsheet.

Signed-off-by: Aloisio Almeida Jr <aloisio.almeida@openbossa.org>